requirements.txt
streamlit==1.35.0
google-generativeai==0.3.2
python-dotenv==1.0.1
catboost==1.2.5
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
